+++ src/cppNGSD/GenLabDB.h
@@ -8,6 +8,13 @@
 #include "PhenotypeList.h"
 #include "NGSD.h"
 
+//accounting data
+struct AccountingData
+{
+	QString insurance_company;
+	QString accounting_mode;
+};
+
 /// GenLab database access (only views that serve as API)
 class CPPNGSDSHARED_EXPORT GenLabDB
 		: public QObject
@@ -88,6 +95,9 @@ public:
 
 	///Returns the tissue type of a sample, or an empty string if it could not be determined.
 	QString tissue(QString ps_name);
+
+	///Returns the accounting data.
+	AccountingData accountingData(QString ps_name);
 
 protected:
 	///Copy constructor "declared away".