+++ doc/GSvar/changelog.md
@@ -1,5 +1,14 @@
 # GSvar change log
 
+## 2025_05-32 (01.07.2025)
+
+- germline report: added HGNC ID to gene table.
+- 
+
+## 2025_05-16 (11.06.2025)
+
+- added support for new megSAP pipeline with separate DRAGEN and megSAP analysis
+
 ## 2025_03-80 (20.05.25)
 
 - RE search: added order date and patient identifier to result table